passing user data to frontend

+++ controllers/blogController.js
@@ -1,6 +1,7 @@
 const myBlogModel = require("../models/blogModel");
 const cloudinary = require("cloudinary");
 const { isEmpty } = require("../common/common");
+const jsonwebtoken = require("jsonwebtoken");
 
 cloudinary.config({
   cloud_name: "abhi-jain",
@@ -34,24 +35,38 @@ function fetchSingleBlog(req, res) {
 
 function userSpecificBlogs(req, res) {
   const userid = req.query;
-  // userid = {id: "dfsrfef"}; this id will refer to user id.
-  if (Object.hasOwnProperty(userid, "id") === false && userid.id.length === 0) {
-    return res
-      .status(404)
-      .send({ message: "You forgot to tell which user's blog you want" });
-  }
-  myBlogModel
-    .find({
-      userId: userid.id,
-    })
-    .then((data) => {
-      return res
-        .status(200)
-        .send({ message: "Blog fetched successfully", blogData: data });
-    })
-    .catch(() => {
-      return res.status(500).send({ message: "Server failed to send blogs" });
-    });
+  jsonwebtoken.verify(req.token, process.env.SECRETKEY, (error, result) => {
+    console.log(error);
+    if (error !== null) {
+      return res.status(404).send({ message: "Token verification failed." });
+    }
+
+    if (result !== undefined) {
+      // userid = {id: "dfsrfef"}; this id will refer to user id.
+      if (
+        Object.hasOwnProperty(userid, "id") === false &&
+        userid.id.length === 0
+      ) {
+        return res
+          .status(404)
+          .send({ message: "You forgot to tell which user's blog you want" });
+      }
+      myBlogModel
+        .find({
+          userId: userid.id,
+        })
+        .then((data) => {
+          return res
+            .status(200)
+            .send({ message: "Blog fetched successfully", blogData: data });
+        })
+        .catch(() => {
+          return res
+            .status(500)
+            .send({ message: "Server failed to send blogs" });
+        });
+    }
+  });
 }
 
 // app.delete("/deleteBlog", (req, res) => );
@@ -72,21 +87,35 @@ function getAllBlogs(req, res) {
 
 function deleteBlog(req, res) {
   const blogId = req.query;
-  // blogId = {id: "dfsrfef"};
-  if (Object.hasOwnProperty(blogId, "id") === false && blogId.id.length === 0) {
-    return res
-      .status(404)
-      .send({ message: "You forgot to tell which blog to delete." });
-  }
+  jsonwebtoken.verify(req.token, process.env.SECRETKEY, (error, result) => {
+    console.log(error);
+    if (error !== null) {
+      return res.status(404).send({ message: "Token verification failed." });
+    }
 
-  myBlogModel
-    .findByIdAndDelete(blogId.id)
-    .then(() => {
-      return res.status(200).send({ message: "Blog deleted successfully." });
-    })
-    .catch(() => {
-      return res.status(404).send({ message: "Failed to delete Blog" });
-    });
+    if (result !== undefined) {
+      // blogId = {id: "dfsrfef"};
+      if (
+        Object.hasOwnProperty(blogId, "id") === false &&
+        blogId.id.length === 0
+      ) {
+        return res
+          .status(404)
+          .send({ message: "You forgot to tell which blog to delete." });
+      }
+
+      myBlogModel
+        .findByIdAndDelete(blogId.id)
+        .then(() => {
+          return res
+            .status(200)
+            .send({ message: "Blog deleted successfully." });
+        })
+        .catch(() => {
+          return res.status(404).send({ message: "Failed to delete Blog" });
+        });
+    }
+  });
 }
 
 // app.put("/updateBlog", upload.single("blogImage"), (req, res) => );
@@ -96,62 +125,79 @@ function updateBlog(req, res) {
   const body = req.body;
   console.log(body);
   const blogId = req.query;
-  // blogId = {id: "dfsrfef"};
-  if (Object.hasOwnProperty(blogId, "id") === false && blogId.id.length === 0) {
-    return res
-      .status(404)
-      .send({ message: "You forgot to tell which blog to update." });
-  }
-  if (blogImage !== undefined) {
-    cloudinary.v2.uploader
-      .upload_stream((error, result) => {
-        if (error !== undefined) {
-          return res
-            .status(500)
-            .send({ message: "Server failed to update image." });
-        }
+  jsonwebtoken.verify(req.token, process.env.SECRETKEY, (error, result) => {
+    console.log(error);
+    if (error !== null) {
+      return res.status(404).send({ message: "Token verification failed." });
+    }
 
-        if (isEmpty(body) === true) {
-          myBlogModel
-            .findByIdAndUpdate(blogId.id, { blogImage: result.url, ...body })
-            .then(() => {
-              return res
-                .status(201)
-                .send({ message: "Blog updated successfully" });
-            })
-            .catch(() => {
-              return res
-                .status(404)
-                .send({ message: "Required Blog does not found." });
-            });
-        } else {
-          myBlogModel
-            .findByIdAndUpdate(blogId.id, { blogImage: result.url })
-            .then(() => {
-              return res
-                .status(201)
-                .send({ message: "Blog updated successfully" });
-            })
-            .catch(() => {
-              return res
-                .status(404)
-                .send({ message: "Required Blog does not found." });
-            });
-        }
-      })
-      .end(blogImage.buffer);
-  } else {
-    myBlogModel
-      .findByIdAndUpdate(blogId.id, { ...body })
-      .then(() => {
-        return res.status(201).send({ message: "Blog updated successfully" });
-      })
-      .catch(() => {
+    if (result !== undefined) {
+      // blogId = {id: "dfsrfef"};
+      if (
+        Object.hasOwnProperty(blogId, "id") === false &&
+        blogId.id.length === 0
+      ) {
         return res
           .status(404)
-          .send({ message: "Required Blog does not found." });
-      });
-  }
+          .send({ message: "You forgot to tell which blog to update." });
+      }
+      if (blogImage !== undefined) {
+        cloudinary.v2.uploader
+          .upload_stream((error, result) => {
+            if (error !== undefined) {
+              return res
+                .status(500)
+                .send({ message: "Server failed to update image." });
+            }
+
+            if (isEmpty(body) === true) {
+              myBlogModel
+                .findByIdAndUpdate(blogId.id, {
+                  blogImage: result.url,
+                  ...body,
+                })
+                .then(() => {
+                  return res
+                    .status(201)
+                    .send({ message: "Blog updated successfully" });
+                })
+                .catch(() => {
+                  return res
+                    .status(404)
+                    .send({ message: "Required Blog does not found." });
+                });
+            } else {
+              myBlogModel
+                .findByIdAndUpdate(blogId.id, { blogImage: result.url })
+                .then(() => {
+                  return res
+                    .status(201)
+                    .send({ message: "Blog updated successfully" });
+                })
+                .catch(() => {
+                  return res
+                    .status(404)
+                    .send({ message: "Required Blog does not found." });
+                });
+            }
+          })
+          .end(blogImage.buffer);
+      } else {
+        myBlogModel
+          .findByIdAndUpdate(blogId.id, { ...body })
+          .then(() => {
+            return res
+              .status(201)
+              .send({ message: "Blog updated successfully" });
+          })
+          .catch(() => {
+            return res
+              .status(404)
+              .send({ message: "Required Blog does not found." });
+          });
+      }
+    }
+  });
 }
 
 function newBlog(req, res) {
@@ -162,32 +208,41 @@ function newBlog(req, res) {
     return res.status(404).send({ message: "You forgot to include image." });
   }
 
-  cloudinary.v2.uploader
-    .upload_stream((error, result) => {
-      if (error !== undefined) {
-        return res
-          .status(500)
-          .send({ message: "Server failed to upload image." });
-      }
+  jsonwebtoken.verify(req.token, process.env.SECRETKEY, (error, result) => {
+    console.log(error);
+    if (error !== null) {
+      return res.status(404).send({ message: "Token verification failed." });
+    }
 
-      const newBlog = myBlogModel({
-        blogImage: result.url,
-        ...body,
-      });
-      newBlog
-        .save()
-        .then(() => {
-          return res
-            .status(201)
-            .send({ message: "Congratulations for new blog." });
+    if (result !== undefined) {
+      cloudinary.v2.uploader
+        .upload_stream((error, result) => {
+          if (error !== undefined) {
+            return res
+              .status(500)
+              .send({ message: "Server failed to upload image." });
+          }
+
+          const newBlog = myBlogModel({
+            blogImage: result.url,
+            ...body,
+          });
+          newBlog
+            .save()
+            .then(() => {
+              return res
+                .status(201)
+                .send({ message: "Congratulations for new blog." });
+            })
+            .catch(() => {
+              return res
+                .status(500)
+                .send({ message: "Server failed to save blog." });
+            });
         })
-        .catch(() => {
-          return res
-            .status(500)
-            .send({ message: "Server failed to save blog." });
-        });
-    })
-    .end(blogImage.buffer);
+        .end(blogImage.buffer);
+    }
+  });
 }
 
 // app.post("/newBlog", upload.single("blogImage"), (req, res) => );
